+++ tests/test_hello.py
@@ -1,5 +1,0 @@
-import onur.main
-
-
-def test_hello():
-    assert onur.main.welcome() == "Hello, World!"